+++ requirements.in
@@ -1,4 +1,4 @@
-git+ssh://git@github.com/codecov/shared.git@680951c4849074db131353b251581ee03029a72a#egg=shared
+git+ssh://git@github.com/codecov/shared.git@96376263de4d4684d7b7d06c2d612f340eab5611#egg=shared
 git+ssh://git@github.com/codecov/opentelem-python.git@v0.0.4a1#egg=codecovopentelem
 boto3
 celery
