+++ requirements.txt
@@ -251,7 +251,7 @@ s3transfer==0.3.4
     # via boto3
 sentry-sdk==1.19.1
     # via -r requirements.in
-shared @ git+ssh://git@github.com/codecov/shared.git@680951c4849074db131353b251581ee03029a72a
+shared @ git+ssh://git@github.com/codecov/shared.git@96376263de4d4684d7b7d06c2d612f340eab5611
     # via -r requirements.in
 six==1.15.0
     # via
