+++ services/owner.py
@@ -3,7 +3,7 @@ import logging
 import shared.torngit as torngit
 from shared.config import get_config, get_verify_ssl
 
-from database.models import Owner
+from helpers.token_refresh import get_token_refresh_callback
 from services.bots import get_owner_appropriate_bot_token
 
 log = logging.getLogger(__name__)
@@ -26,6 +26,11 @@ def get_owner_provider_service(owner, using_integration=False):
         oauth_consumer_token=dict(
             key=get_config(service, "client_id"),
             secret=get_config(service, "client_secret"),
+        ),
+        # if using integration we will use the integration token
+        # not the owner's token
+        on_token_refresh=(
+            get_token_refresh_callback(owner) if not using_integration else None
         ),
     )
     return _get_owner_provider_service_instance(service, **adapter_params)
